backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.6
bcrypt==4.1.2